log enqueue

+++ Plugins/RawCMS.Plugins.LogCollecting/Model/QueueLoad.cs
@@ -6,7 +6,7 @@ namespace RawCMS.Plugins.LogCollecting.Models
     public class QueueLoad
     {
         public DateTime Time { get; set; }
-        public int QueueSize { get; set; }
-        public int MaxSize { get; set; }
+        public long QueueSize { get; set; }
+        public long MaxSize { get; set; }
     }
 }
